export Endo newtype wrapper
